Print detailed computational differences.

+++ src/cn/edu/sjtu/ddst/fptest/Tester.java
@@ -3,6 +3,7 @@ package cn.edu.sjtu.ddst.fptest;
 import java.io.*;
 import java.nio.file.Path;
 import java.text.SimpleDateFormat;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Date;
 
@@ -14,24 +15,24 @@ public class Tester {
     public static final int TEST_ALL = TEST_OPERATOR | TEST_MATH_FUNC;
 
     private int range;
-    private long seed;
-    private int length;
+    private long genSeed;
+    private int programLen;
 
     public Tester(int testRange, int programLength, long generatorSeed) {
         this.range = testRange;
-        this.length = programLength;
-        this.seed = generatorSeed;
+        this.programLen = programLength;
+        this.genSeed = generatorSeed;
     }
 
     private static final String OUTPUT_DIR = "./out";
 
-    public void test(TestConfig[] configs, int rounds) {
+    public void test(TestConfig[] configs, int nRounds) {
         // Initialize generator list
         if (configs.length == 0)
             throw new RuntimeException("Must specify at least one testing configuration");
         Generator[] generators = new Generator[configs.length];
         for (int i = 0; i < configs.length; i++)
-            generators[i] = new Generator(seed);
+            generators[i] = new Generator(genSeed);
 
         // Create output directory
         Path outPath = Path.of(OUTPUT_DIR);
@@ -55,6 +56,7 @@ public class Tester {
             writer = new FileWriter(logPath.toString());
             for (TestConfig config : configs)
                 writer.write(config.toString() + '\n');
+            writer.write('\n');
             writer.flush();
         } catch (IOException e) {
             e.printStackTrace();
@@ -63,46 +65,43 @@ public class Tester {
 
         // Test for specified number of rounds
         int diffCount = 0;
-        for (int i = 0; i < rounds; i++) {
-            System.out.printf("\rProgress: %d/%d\r", i, rounds);
+        for (int iRound = 0; iRound < nRounds; iRound++) {
+            System.out.printf("\rProgress: %d/%d\r", iRound, nRounds);
 
             // Initialize result array
-            String[] outputs = new String[configs.length];
+            String[][] outputs = new String[configs.length][programLen + 1];
 
             // Generate and run according to specified configuration
-            for (int j = 0; j < configs.length; j++) {
-                String name = String.format("Test%dConfig%d", i, j);
-                Program program = generators[j].generate(length, range, configs[j].strict, name);
-                String result = runOneProgram(program, testDir, configs[j]);
-                if (result.isEmpty())
-                    throw new RuntimeException("Empty output from test program");
-                outputs[j] = result;
+            for (int iConfig = 0; iConfig < configs.length; iConfig++) {
+                String name = String.format("Test%dConfig%d", iRound, iConfig);
+                Program program = generators[iConfig].generate(programLen, range,
+                        configs[iConfig].strict, name);
+                outputs[iConfig] = runOneProgram(program, testDir, configs[iConfig]);
             }
 
             // Judge whether outputs of programs are all the same
-            boolean equals = true;
-            for (int j = 1; j < configs.length; j++)
-                equals &= outputs[j].equals(outputs[0]);
-            if (!equals) {
-                // Print difference in long bits
-                diffCount++;
-                double[] fpRes = new double[configs.length];
-                for (int j = 0; j < configs.length; j++)
-                    fpRes[j] = Double.longBitsToDouble(Long.parseUnsignedLong(outputs[j], 16));
-                String diffStr = String.format("Difference detected in round %d: %s (%s)\n",
-                        i, Arrays.toString(outputs), Arrays.toString(fpRes));
-                System.out.print(diffStr);
-                try {
-                    writer.write(diffStr);
-                    writer.flush();
-                } catch (IOException e) {
-                    System.err.println("Cannot write to log file " + logPath.toString());
-                    e.printStackTrace();
-                }
+            boolean allEquals = true;
+            for (int iConfig = 1; iConfig < configs.length; iConfig++)
+                allEquals &= Arrays.equals(outputs[0], outputs[iConfig]);
+            if (allEquals) continue;
+            diffCount++;
+
+            // Record difference
+            String detectStr = String.format("Difference detected in round %d.\n", iRound);
+            System.out.print(detectStr);
+            try {
+                writer.write(detectStr);
+                writer.write("Details: \n");
+                writer.write(printDetailedDifference(outputs));
+                writer.write('\n');
+                writer.flush();
+            } catch (IOException e) {
+                e.printStackTrace();
+                throw new RuntimeException("Cannot write difference details.");
             }
         }
 
-        System.out.printf("Count: %d/%d\n", diffCount, rounds);
+        System.out.printf("Count: %d/%d\n", diffCount, nRounds);
         try {
             writer.close();
         } catch (IOException e) {
@@ -110,7 +109,7 @@ public class Tester {
         }
     }
 
-    private static String runOneProgram(Program program, Path testDir, TestConfig config) {
+    private String[] runOneProgram(Program program, Path testDir, TestConfig config) {
         // Write source code to file
         String srcName = program.name + ".java";
         File srcFile = testDir.resolve(srcName).toFile();
@@ -141,7 +140,7 @@ public class Tester {
 
         // Run compiled class in specified JVM
         Path jvmPath =  Path.of(config.jreHome, "bin" , "java");
-        StringBuilder output = new StringBuilder();
+        ArrayList<String> lines = new ArrayList<>(programLen + 1);
         try {
             String command = String.format("%s %s %s", jvmPath.toString(), config.runOptions,
                     program.name);
@@ -149,16 +148,52 @@ public class Tester {
             BufferedReader reader = new BufferedReader(new InputStreamReader(proc.getInputStream()));
             String line;
             while ((line = reader.readLine()) != null)
-                output.append(line);
+                lines.add(line);
         } catch (IOException e) {
             e.printStackTrace();
             throw new RuntimeException("Failed to run compiled class.");
         }
 
+        // Check if output lines are valid
+        if (lines.size() != programLen + 1)
+            throw new RuntimeException(
+                    String.format("Expect %d lines of outout, got %d.", programLen + 1, lines.size())
+            );
+        String[] out = new String[programLen + 1];
+        lines.toArray(out);
+
         // Clean up
         if (!testDir.resolve(program.name + ".class").toFile().delete())
             System.err.println("Failed to delete class file");
 
-        return output.toString();
+        return out;
+    }
+
+    private static String printDetailedDifference(String[][] outputs) {
+        // Find where the difference begins
+        boolean[] configEq = new boolean[outputs[0].length];
+        int begin = -1;
+        for (int iState = 0; iState < outputs[0].length; iState++) {
+            boolean allEq = true;
+            for (int iConfig = 1; iConfig < outputs.length; iConfig++) {
+                allEq &= outputs[0][iState].equals(outputs[iConfig][iState]);
+            }
+            configEq[iState] = allEq;
+            if (!allEq && begin < 0) begin = iState;
+        }
+
+        // Tabulate all the results from the where the difference begins
+        StringBuilder builder = new StringBuilder();
+        for (int iState = Math.max(0, begin - 1); iState < outputs[0].length; iState++) {
+            StringBuilder line = new StringBuilder(String.format("%1$8d\t", iState));
+            for (String[] output : outputs) {
+                String result = output[iState];
+                double value = Double.longBitsToDouble(Long.parseUnsignedLong(result, 16));
+                line.append(String.format("%1$16s(%2$24s)\t", result, value));
+            }
+            builder.append(configEq[iState] ? ' ' : '>').append(line).append('\n');
+        }
+
+        return builder.toString();
     }
 }
